requirements.txt
tensorflow
keras
h5py
Pillow
numpy
oauth2client
google-api-python-client
scikit-learn
requests
google-cloud-storage